requirements.txt
azure-mgmt-network
azure-identity